Add user score

+++ polling_application/src/main/java/com/example/polling_application/repositories/QuestionRepository.java
@@ -1,7 +1,11 @@
 package com.example.polling_application.repositories;
 
 import com.example.polling_application.entities.Question;
+import com.example.polling_application.entities.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface QuestionRepository extends JpaRepository<Question,Long> {
+import java.util.List;
+
+public interface QuestionRepository extends JpaRepository<Question,Integer> {
+
 }
